requirements.txt
streamlit>=1.43.0
selenium>=4.0.0
beautifulsoup4>=4.9.0
pandas>=1.3.0
plotly>=6.0.0
webdriver-manager>=3.8.0
requests-html>=0.10.0
lxml==4.9.3
